abstracted out keyboard and window
